chore: restore auth define-config
