+++ src/libs/better-auth/define-config.ts
@@ -1,16 +1,20 @@
 /* eslint-disable sort-keys-fix/sort-keys-fix, typescript-sort-keys/interface */
+import { expo } from '@better-auth/expo';
+import { passkey } from '@better-auth/passkey';
 import { createNanoId, idGenerator, serverDB } from '@lobechat/database';
+import * as schema from '@lobechat/database/schemas';
 import { emailHarmony } from 'better-auth-harmony';
 import { drizzleAdapter } from 'better-auth/adapters/drizzle';
 import { BetterAuthOptions, betterAuth } from 'better-auth/minimal';
-import { admin, genericOAuth, magicLink } from 'better-auth/plugins';
-import type { BetterAuthPlugin } from 'better-auth/types';
+import { admin, emailOTP, genericOAuth, magicLink } from 'better-auth/plugins';
+import { BetterAuthPlugin } from 'better-auth/types';
 
 import { authEnv } from '@/envs/auth';
 import {
   getMagicLinkEmailTemplate,
   getResetPasswordEmailTemplate,
   getVerificationEmailTemplate,
+  getVerificationOTPEmailTemplate,
 } from '@/libs/better-auth/email-templates';
 import { initBetterAuthSSOProviders } from '@/libs/better-auth/sso';
 import { createSecondaryStorage, getTrustedOrigins } from '@/libs/better-auth/utils/config';
@@ -21,7 +25,34 @@ import { UserService } from '@/server/services/user';
 // Email verification link expiration time (in seconds)
 // Default is 1 hour (3600 seconds) as per Better Auth documentation
 const VERIFICATION_LINK_EXPIRES_IN = 3600;
+
+/**
+ * Safely extract hostname from AUTH_URL for passkey rpID.
+ * Returns undefined if AUTH_URL is not set (e.g., in e2e tests).
+ */
+const getPasskeyRpID = (): string | undefined => {
+  if (!authEnv.NEXT_PUBLIC_AUTH_URL) return undefined;
+  try {
+    return new URL(authEnv.NEXT_PUBLIC_AUTH_URL).hostname;
+  } catch {
+    return undefined;
+  }
+};
+
+/**
+ * Get passkey origins array.
+ * Returns undefined if AUTH_URL is not set (e.g., in e2e tests).
+ */
+const getPasskeyOrigins = (): string[] | undefined => {
+  if (!authEnv.NEXT_PUBLIC_AUTH_URL) return undefined;
+  return [
+    // Web origin
+    authEnv.NEXT_PUBLIC_AUTH_URL,
+  ];
+};
 const MAGIC_LINK_EXPIRES_IN = 900;
+// OTP expiration time (in seconds) - 5 minutes for mobile OTP verification
+const OTP_EXPIRES_IN = 300;
 const enableMagicLink = authEnv.NEXT_PUBLIC_ENABLE_MAGIC_LINK;
 const enabledSSOProviders = parseSSOProviders(authEnv.AUTH_SSO_PROVIDERS);
 
@@ -91,6 +122,8 @@ export function defineConfig(customOptions: CustomBetterAuthOptions) {
     },
     database: drizzleAdapter(serverDB, {
       provider: 'pg',
+      // experimental joins feature needs schema to pass full relation
+      schema,
     }),
     secondaryStorage: createSecondaryStorage(),
     /**
@@ -156,8 +189,43 @@ export function defineConfig(customOptions: CustomBetterAuthOptions) {
     },
     plugins: [
       ...customOptions.plugins,
+      expo(),
       emailHarmony({ allowNormalizedSignin: false }),
       admin(),
+      // Email OTP plugin for mobile verification
+      emailOTP({
+        expiresIn: OTP_EXPIRES_IN,
+        otpLength: 6,
+        allowedAttempts: 3,
+        // Don't automatically send OTP on sign up - let mobile client manually trigger it
+        sendVerificationOnSignUp: false,
+        async sendVerificationOTP({ email, otp }) {
+          const emailService = new EmailService();
+
+          // For all OTP types, use the same template
+          // userName is optional and will be null since we don't have user context here
+          const template = getVerificationOTPEmailTemplate({
+            expiresInSeconds: OTP_EXPIRES_IN,
+            otp,
+            userName: null,
+          });
+
+          await emailService.sendMail({
+            to: email,
+            ...template,
+          });
+        },
+      }),
+      passkey({
+        rpName: 'LobeHub',
+        // Extract rpID from auth URL (e.g., 'lobehub.com' from 'https://lobehub.com')
+        // Returns undefined if AUTH_URL is not set (e.g., in e2e tests)
+        rpID: getPasskeyRpID(),
+        // Support multiple origins: web + Android APK key hashes
+        // Android origin format: android:apk-key-hash:<base64url-sha256-fingerprint>
+        // Returns undefined if AUTH_URL is not set (e.g., in e2e tests)
+        origin: getPasskeyOrigins(),
+      }),
       ...(genericOAuthProviders.length > 0
         ? [
             genericOAuth({
